feat: wait for channelData when activated
It's possible that the process() function will be called without any channel data for a few times
when connecting an input. If the recording gets activated before the process() function receives any
input it previously threw an error. It will now wait to be called with channel data before switching
the state to recording.

fix chrisguttandin/extendable-media-recorder-wav-encoder#393