feat: keep the recorder alive in case of missing samples
Previously the recorder stopped when the input given to the process function was empty. This causes
a race condition since there is no way for the code that is intended to stop the recorder to know if
it is save to stop the recorder or if it has been stopped already.

BREAKING CHANGE: Technically this is a breaking change as it changes the behaviour of the recorder
and may cause longer recordings as before.

fix chrisguttandin/extendable-media-recorder#675